Refactor package names in com.ibm.ws.microprofile.metrics.cdi.2.0 and tweak bnd file in the same bundle as well as com.ibm.ws.microprofile.metrics.cdi.common
