+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/bnd.bnd
@@ -14,7 +14,7 @@ bVersion=1.0
 javac.source: 1.8
 javac.target: 1.8
 
-Bundle-Name: com.ibm.ws.microprofile.metrics.cdi
+Bundle-Name: com.ibm.ws.microprofile.metrics.cdi.2.0
 Bundle-SymbolicName: com.ibm.ws.microprofile.metrics.cdi.2.0
 Bundle-Description: MicroProfile Metrics CDI Integration, version ${bVersion}
 
@@ -23,16 +23,16 @@ Import-Package: \
 	*
 
 Export-Package: \
-	io.astefanutti.metrics.cdi,\
+	io.astefanutti.metrics.cdi20,\
 	com.ibm.ws.microprofile.metrics.cdi.decorator,\
 	com.ibm.ws.microprofile.metrics.cdi.binding,\
-	com.ibm.ws.microprofile.metrics.cdi.interceptors,\
-	com.ibm.ws.microprofile.metrics.cdi.producer,\
-	com.ibm.ws.microprofile.metrics.cdi.helper
+	com.ibm.ws.microprofile.metrics.cdi20.interceptors,\
+	com.ibm.ws.microprofile.metrics.cdi20.producer,\
+	com.ibm.ws.microprofile.metrics.cdi20.helper
 
 Private-Package: \
-    com.ibm.ws.microprofile.metrics.cdi.resources,\
-    com.ibm.ws.microprofile.metrics.cdi.helper
+	com.ibm.ws.microprofile.metrics.cdi.resources,\
+	com.ibm.ws.microprofile.metrics.cdi20.helper
 
 Include-Resource: \
     META-INF=resources/META-INF
@@ -41,7 +41,7 @@ WS-TraceGroup: METRICS
 
 -dsannotations-inherit: true
 -dsannotations: \
-	io.astefanutti.metrics.cdi.MetricsExtension
+	io.astefanutti.metrics.cdi20.MetricsExtension
 
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
