+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-io.astefanutti.metrics.cdi.MetricsExtension
+io.astefanutti.metrics.cdi20.MetricsExtension
