+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/src/com/ibm/ws/microprofile/metrics/cdi20/helper/package-info.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2019 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -12,4 +12,4 @@
  * @version 1.0
  */
 @org.osgi.annotation.versioning.Version("1.0")
-package com.ibm.ws.microprofile.metrics.cdi.producer;
+package com.ibm.ws.microprofile.metrics.cdi20.helper;
