+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/src/com/ibm/ws/microprofile/metrics/cdi20/interceptors/package-info.java
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 /**
- * @version 1.0
+ *
  */
 @org.osgi.annotation.versioning.Version("1.0")
-package com.ibm.ws.microprofile.metrics.cdi.helper;
+package com.ibm.ws.microprofile.metrics.cdi20.interceptors;
