+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/src/com/ibm/ws/microprofile/metrics/cdi20/producer/MetricRegistryFactory.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.microprofile.metrics.cdi.producer;
+package com.ibm.ws.microprofile.metrics.cdi20.producer;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Produces;
