+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/bnd.bnd
@@ -16,7 +16,7 @@ javac.target: 1.8
 
 Bundle-Name: com.ibm.ws.microprofile.metrics.cdi.common
 Bundle-SymbolicName: com.ibm.ws.microprofile.metrics.cdi.common
-Bundle-Description: MicroProfile Metrics CDI Integration, version ${bVersion}
+Bundle-Description: MicroProfile Metrics CDI common bundle, version ${bVersion}
 
 Import-Package: \
 	javax.enterprise.*;version="[1.1,3)",\
@@ -26,11 +26,11 @@ Export-Package: \
     com.ibm.ws.microprofile.metrics.cdi.decorator,\
     com.ibm.ws.microprofile.metrics.cdi.binding
 
-Private-Package: \
-    com.ibm.ws.microprofile.metrics.cdi.resources
+#Private-Package: \
+#    com.ibm.ws.microprofile.metrics.cdi.resources
 
-Include-Resource: \
-    META-INF=resources/META-INF
+#Include-Resource: \
+#    META-INF=resources/META-INF
 
 WS-TraceGroup: METRICS
 publish.wlp.jar.disabled: true
